resume_ats_scanner/backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PDF
pdfplumber>=0.10.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0

# NLP & ML
spacy>=3.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Fuzzy matching
rapidfuzz>=3.5.0

# PDF Report
reportlab>=4.0.0

# Database
pymongo>=4.5.0

# Optional - AI Enhancement
openai>=1.0.0

# Utils
python-dotenv>=1.0.0
